fastapi_task_manager/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
SQLAlchemy>=1.4
pydantic>=1.10
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.0.1
pytest>=7.0
httpx>=0.23
